renamed core.ts => note-graph.ts
