+++ packages/foam-core/src/index.ts
@@ -1,2 +1,2 @@
-export { NoteGraph, Note, NoteLink } from './core'
+export { NoteGraph, Note, NoteLink } from './note-graph'
 export { createNoteFromMarkdown, createMarkdownReferences } from './markdown-provider'
